make StreamWrapper independent from extensions
StreamWrapper acts as a base class for classes that implement stream extension
methods. RxStream is such a sub class. It does not contain any extension
methods yet. StreamWrapperType is used as a mixin to provide concrete return
types to some of the methods that sub classes like RxStream inherit from
StreamWrapper.